requirements.txt
python-dotenv
kafka-python
pyspark
numpy
transformers
elasticsearch
sentence-transformers